Changed result tensor type from gt op

+++ aten/src/ATen/native/LinearAlgebra.cpp
@@ -126,7 +126,7 @@ Tensor matrix_rank(const Tensor& self, double tol, bool symmetric) {
            "of floating types");
 
   Tensor S = _matrix_rank_helper(self, symmetric);
-  return (S > tol).sum();
+  return (S > tol).to(at::kByte).sum();
 }
 
 Tensor matrix_rank(const Tensor& self, bool symmetric) {
@@ -136,7 +136,7 @@ Tensor matrix_rank(const Tensor& self, bool symmetric) {
 
   Tensor S = _matrix_rank_helper(self, symmetric);
   double tol = _get_epsilon(self.scalar_type()) * std::max(self.size(0), self.size(1));
-  return (S > S.max().mul_(tol)).sum();
+  return (S > S.max().mul_(tol)).to(at::kByte).sum();
 }
 
 static void check_1d(const Tensor& t, const char* arg, const char* fn) {
